build: warning triage for the viz+tests configuration
CI's test lane builds visualization together with the test suite — a
combination the first triage pass missed locally. Same finding classes
in the decomp visual/test headers and the sampling planners: init-order,
unused interface-default parameters, and sign-compare on size()-1 loop
bounds (the casts also close the empty-container wraparound those
bounds carried). Verified: viz-on 23/23, viz-off + ASan/UBSan 14/14.

+++ src/planning/decomp/include/decomp/details/dense_grid_visual.hpp
@@ -70,7 +70,7 @@ void ShowPathOnMatrixAsColorMap(const Eigen::MatrixXd &matrix,
 
   quickviz::CvCanvas canvas(color_img);
 
-  for (int i = 0; i < waypoints.size() - 1; ++i) {
+  for (int i = 0; i < static_cast<int>(waypoints.size()) - 1; ++i) {
     canvas.DrawLine({static_cast<double>(waypoints[i].GetX()),
                      static_cast<double>(waypoints[i].GetY())},
                     {static_cast<double>(waypoints[i + 1].GetX()),
